perf: accept new cart for applying discounts
BREAKING CHANGE: The functions used for applying discounts to a cart
now accept the new cart, rather that the cart id. This is required
because when calculating the discounts previously, the items of the persisted
cart were used, rather than the items of the new cart that is in the process of
being persisted.

Signed-off-by: tedraykov <tedraykov@gmail.com>